requirements-dev.txt
# Full requirements for developments
pytorch-ignite==0.4.8
PyWavelets==1.3.0

scipy==1.8.1
#itk>=5.2
nibabel==3.2.2
pillow!=8.3.0  # https://github.com/python-pillow/Pillow/issues/5571
tensorboard==2.9.0
scikit-image==0.19.2
tqdm>=4.47.0
lmdb==1.3.0
flake8>=3.8.1
flake8-bugbear
flake8-comprehensions
flake8-executable
flake8-pyi
pylint!=2.13  # https://github.com/PyCQA/pylint/issues/5969
mccabe
pep8-naming
pycodestyle
pyflakes==2.4.0
black
isort
pytype>=2020.6.1; platform_system != "Windows"
types-pkg_resources
mypy>=0.790
ninja==1.10.2.3
torchvision==0.12.0
opencv-python==4.5.5.64
psutil==5.9.1
Sphinx==3.5.3
recommonmark==0.6.0
sphinx-autodoc-typehints==1.11.1
sphinx-rtd-theme==0.5.2
cucim==22.2.1; platform_system == "Linux"
imagecodecs; platform_system == "Linux"
tifffile; platform_system == "Linux"
pandas==1.4.2
requests==2.27.1
einops==0.4.1
transformers==4.19.2
mlflow==1.26.1
matplotlib!=3.5.0
tensorboardX==2.5
types-PyYAML
pyyaml==6.0
fire==0.4.0
jsonschema==4.6.0
# pyradiomics== 3.0.1
monai[nibabel, skimage, pillow, tensorboard, gdown, ignite, torchvision, itk, tqdm, lmdb, psutil, cucim, pandas, einops, transformers, mlflow, matplotlib, tensorboardX, tifffile, imagecodecs]==0.8.1
kymatio==0.2.1
juliacall==0.8.0
jupyterlab_github==3.0.1
torchio==0.18.77
pytorch-lightning==1.6.4
git+https://github.com/sat28/githubcommit.git
gdown==4.4.0
seaborn==0.11.2
PyTorchUtils==0.0.3
optuna==2.10.0
jupyterlab-git==0.37.1
comet-ml==3.31.3
voxelmorph==0.2
tensorflow==2.9.1
tensorflow_addons==0.17.0
ipywidgets==7.7.0
h5py==3.7.0
itk-elastix==0.14.1




# zenodo_get==1.3.4
# ipykernel
# jupyter
#setuptools
